Move homeView to its own folder
