--- src/app/home.view.html
@@ -1,4 +1,0 @@
-home
-
-<button *ngIf="! authService.isAuthenticated()" (click)="login()">Login</button>
-<button *ngIf="authService.isAuthenticated()" (click)="logout()">Logout</button>
