+++ src/app/home/home.view.html
@@ -1,0 +1,5 @@
+<h2>Home View</h2>
+<hr>
+<p>
+	You can navigate to: <a router [routerLink]="['../login']">Login</a>
+</p>
